Layout renamed, added to App
