+++ src/Components/Layout/PublicLayout.js
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../Header/Header";
+import Footer from "../Footer/Footer";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import RegisterForm from "../Auth/RegisterForm";
